+++ guillotina/api/ws.py
@@ -26,7 +26,7 @@ from urllib import parse
 
 @configure.service(
     context=IContainer, method='GET',
-    permission='guillotina.AccessContent', name='@wstoken',
+    permission='guillotina.UseWebSockets', name='@wstoken',
     summary='Return a web socket token',
     responses={
         "200": {
@@ -43,7 +43,7 @@ from urllib import parse
     })
 @configure.service(
     context=IApplication, method='GET',
-    permission='guillotina.AccessContent', name='@wstoken',
+    permission='guillotina.UseWebSockets', name='@wstoken',
     summary='Return a web socket token',
     responses={
         "200": {