+++ guillotina/tests/test_contentapi.py
@@ -10,7 +10,6 @@ async def test_contentapi_create(db, guillotina_main):
         await api.create({'@type': 'Item', 'id': 'foobar'}, in_=container)
         item = await api.get('foobar', in_=container)
         assert get_content_path(item) == '/foobar'
-        await api.delete(container)
 
 
 @pytest.mark.flaky(reruns=5)
@@ -22,4 +21,3 @@ async def test_contentapi_delete(db, guillotina_main):
         assert get_content_path(item) == '/foobar'
         await api.delete(item)
         assert await api.get('foobar', in_=container) is None
-        await api.delete(container)