+++ CHANGELOG.rst
@@ -1,7 +1,8 @@
 4.5.10 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Fix jwt implementation to use customized algorithm for encoding
+  [vangheem]
 
 
 4.5.9 (2019-03-18)