+++ CHANGELOG.rst
@@ -17,6 +17,9 @@
 - Lazy load dynamically generated jwk key
   [vangheem]
 
+- Fix use of default factory for default value on content
+  [vangheem]
+
 - Fix jwt implementation to use customized algorithm for encoding
   [vangheem]